Summary: Merge upstream code

Test Plan: CI pipeline


Issue: https://github.com/alibaba/dragonwell8/issues/301

+++ hotspot/src/share/vm/opto/loopnode.hpp
@@ -350,7 +350,7 @@ public:
 
   Node_List _body;              // Loop body for inner loops
 
-  uint8 _nest;                  // Nesting depth
+  uint16_t _nest;               // Nesting depth
   uint8 _irreducible:1,         // True if irreducible
         _has_call:1,            // True if has call safepoint
         _has_sfpt:1,            // True if has non-call safepoint
